modify

+++ frameworks/ets/ani/src/sts_action_button.cpp
@@ -17,6 +17,7 @@
 #include "sts_common.h"
 #include "sts_convert_other.h"
 #include "ani_common_want.h"
+#include "ans_log_wrapper.h"
 #include "sts_user_input.h"
 
 
@@ -24,79 +25,114 @@ namespace OHOS {
 namespace NotificationSts {
 void GetStsActionButtonByOther(StsActionButton &actionButton)
 {
+    ANS_LOGD("GetStsActionButtonByOther call");
     actionButton.icon = nullptr;
     actionButton.semanticActionButton = SemanticActionButton::NONE_ACTION_BUTTON;
     actionButton.autoCreatedReplies = true;
     actionButton.mimeTypeOnlyInputs = {};
     actionButton.isContextual = false;
+    ANS_LOGD("GetStsActionButtonByOther end");
 }
 ani_status GetStsActionButtonByWantAgent(ani_env *env, ani_object param,
     StsActionButton &actionButton)
 {
+    ANS_LOGD("GetStsActionButtonByWantAgent call");
+    if (env == nullptr || param == nullptr) {
+        ANS_LOGE("GetStsActionButtonByWantAgent fail, has nullptr");
+        return ANI_ERROR;
+    }
     ani_boolean isUndefind = ANI_TRUE;
     ani_ref wantAgentRef;
     if (ANI_OK != GetPropertyRef(env, param, "wantAgent", isUndefind, wantAgentRef) || isUndefind == ANI_TRUE) {
+        ANS_LOGE("GetStsActionButtonByWantAgent: GetPropertyRef wantAgent failed");
         deletePoint(wantAgentRef);
         return ANI_INVALID_ARGS;
     }
     std::shared_ptr<WantAgent> wantAgent = UnwrapWantAgent(env, static_cast<ani_object>(wantAgentRef));
     deletePoint(wantAgentRef);
     if (wantAgent == nullptr) {
+        ANS_LOGE("GetStsActionButtonByWantAgent: wantAgent is nullptr");
         return ANI_INVALID_ARGS;
     }
     actionButton.wantAgent = wantAgent;
+    ANS_LOGD("GetStsActionButtonByWantAgent end");
     return ANI_OK;
 }
 
 ani_status GetStsActionButtonByWantParams(ani_env *env, ani_object param,
     StsActionButton &actionButton)
 {
+    ANS_LOGD("GetStsActionButtonByWantParams call");
+    if (env == nullptr || param == nullptr) {
+        ANS_LOGE("GetStsActionButtonByWantParams fail, has nullptr");
+        return ANI_ERROR;
+    }
     ani_boolean isUndefind = ANI_TRUE;
     WantParams wantParams = {};
     ani_ref extrasRef;
     if (ANI_OK == GetPropertyRef(env, param, "extras", isUndefind, extrasRef) && isUndefind == ANI_FALSE) {
         UnwrapWantParams(env, extrasRef, wantParams);
     } else {
+        ANS_LOGE("GetStsActionButtonByWantParams: GetPropertyRef extras failed");
         deletePoint(extrasRef);
         return ANI_INVALID_ARGS;
     }
     std::shared_ptr<WantParams> extras = std::make_shared<WantParams>(wantParams);
     actionButton.extras = extras;
+    ANS_LOGD("GetStsActionButtonByWantParams end");
     return ANI_OK;
 }
 ani_status GetStsActionButtonByUserInput(ani_env *env, ani_object param,
     StsActionButton &actionButton)
 {
+    ANS_LOGD("GetStsActionButtonByUserInput call");
+    if (env == nullptr || param == nullptr) {
+        ANS_LOGE("GetStsActionButtonByUserInput fail, has nullptr");
+        return ANI_ERROR;
+    }
     ani_boolean isUndefind = ANI_TRUE;
     std::shared_ptr<Notification::NotificationUserInput> userInput = nullptr;
     ani_ref userInputRef;
     if(ANI_OK == GetPropertyRef(env, param, "userInput", isUndefind, userInputRef) && isUndefind == ANI_FALSE) {
         UnwrapNotificationUserInput(env, static_cast<ani_object>(userInputRef), userInput);
+    } else {
+        ANS_LOGD("GetStsActionButtonByUserInput : GetPropertyRef userInput failed");
     }
     if (userInput == nullptr) {
+        ANS_LOGD("GetStsActionButtonByUserInput : userInput is nullptr");
         userInput = {};
     }
     actionButton.userInput = userInput;
+    ANS_LOGD("GetStsActionButtonByUserInput end");
     return ANI_OK;
 }
 
 ani_status UnwrapNotificationActionButton(ani_env *env, ani_object param,
     StsActionButton &actionButton)
 {
+    ANS_LOGD("UnwrapNotificationActionButton call");
+    if (env == nullptr || param == nullptr) {
+        ANS_LOGE("UnwrapNotificationActionButton fail, has nullptr");
+        return ANI_ERROR;
+    }
     ani_status status = ANI_ERROR;
     ani_boolean isUndefind = ANI_TRUE;
     std::string title;
     if ((status = GetPropertyString(env, param, "title", isUndefind, title)) != ANI_OK || isUndefind == ANI_TRUE) {
+        ANS_LOGE("UnwrapNotificationActionButton : Get title failed");
         return ANI_INVALID_ARGS;
     }
     actionButton.title = title;
     if (ANI_OK != GetStsActionButtonByWantAgent(env, param, actionButton)) {
+        ANS_LOGE("UnwrapNotificationActionButton : GetStsActionButtonByWantAgent failed");
         return ANI_INVALID_ARGS;
     }
     if (ANI_OK != GetStsActionButtonByWantParams(env, param, actionButton)) {
+        ANS_LOGE("UnwrapNotificationActionButton : GetStsActionButtonByWantParams failed");
         return ANI_INVALID_ARGS;
     }
     GetStsActionButtonByOther(actionButton);
+    ANS_LOGD("UnwrapNotificationActionButton end");
     return status;
 }
 
@@ -106,29 +142,45 @@ bool SetNotificationActionButtonByRequiredParameter(
     ani_object &iconButtonObject,
     const std::shared_ptr<NotificationActionButton> &actionButton)
 {
+    ANS_LOGD("SetActionButtonByRequiredParameter call");
+    if (env == nullptr || iconButtonCls == nullptr || iconButtonObject == nullptr || actionButton == nullptr) {
+        ANS_LOGE("SetActionButtonByRequiredParameter fail, has nullptr");
+        return ANI_ERROR;
+    }
     ani_string stringValue;
     // title: string;
     if (!GetAniStringByString(env, actionButton->GetTitle(), stringValue)) {
         deletePoint(stringValue);
+        ANS_LOGE("SetActionButtonByRequiredParameter: Get title failed");
         return false;
     }
     if (!CallSetter(env, iconButtonCls, iconButtonObject, "title", stringValue)) {
         deletePoint(stringValue);
+        ANS_LOGE("SetActionButtonByRequiredParameter: Set title failed");
         return false;
     }
     // wantAgent: WantAgent;
     std::shared_ptr<AbilityRuntime::WantAgent::WantAgent> agent = actionButton->GetWantAgent();
     if (agent == nullptr) {
-        ANS_LOGI("agent is null");
+        ANS_LOGE("SetActionButtonByRequiredParameter:agent is null");
         deletePoint(stringValue);
         return false;
     } else {
         ani_object wantAgent = AppExecFwk::WrapWantAgent(env, agent.get());
-        if (!CallSetter(env, iconButtonCls, iconButtonObject, "wantAgent", wantAgent)) {
+        if (wantAgent == nullptr) {
+            ANS_LOGE("SetActionButtonByRequiredParameter: wantAgent is nullptr");
             deletePoint(stringValue);
+            deletePoint(wantAgent);
+            return false;
+        }
+        if (!CallSetter(env, iconButtonCls, iconButtonObject, "wantAgent", wantAgent)) {
+            ANS_LOGE("SetActionButtonByRequiredParameter: Set wantAgent failed");
+            deletePoint(stringValue);
+            deletePoint(wantAgent);
             return false;
         }
     }
+    ANS_LOGD("SetActionButtonByRequiredParameter end");
     return true;
 }
 
@@ -138,20 +190,29 @@ void SetNotificationActionButtonByOptionalParameter(
     ani_object &iconButtonObject,
     const std::shared_ptr<NotificationActionButton> &actionButton)
 {
+    ANS_LOGD("SetActionButtonByOptionalParameter call");
+    if (env == nullptr || iconButtonCls == nullptr || iconButtonObject == nullptr || actionButton == nullptr) {
+        ANS_LOGE("SetActionButtonByOptionalParameter fail, has nullptr");
+        return;
+    }
     // extras?: Record<string, Object>
     ani_ref extras = WrapWantParams(env, *(actionButton->GetAdditionalData().get()));
     if (!CallSetter(env, iconButtonCls, iconButtonObject, "extras", extras)) {
+        ANS_LOGD("SetActionButtonByOptionalParameter : Set extras failed");
         deletePoint(extras);
     }
     // userInput?: NotificationUserInput
     ani_object userInputObject = WarpUserInput(env, actionButton->GetUserInput());
     if (!CallSetter(env, iconButtonCls, iconButtonObject, "userInput", userInputObject)) {
+        ANS_LOGD("SetActionButtonByOptionalParameter : Set userInput failed");
         deletePoint(userInputObject);
     }
+    ANS_LOGD("SetActionButtonByOptionalParameter end");
 }
 
 void deletePointOfWrapNotificationActionButton(
     ani_object iconButtonObject, ani_class iconButtonCls, ani_string stringValue) {
+    ANS_LOGD("deletePointOfWrapNotificationActionButton call");
     deletePoint(iconButtonObject);
     deletePoint(iconButtonCls);
     deletePoint(stringValue);
@@ -160,8 +221,9 @@ void deletePointOfWrapNotificationActionButton(
 ani_object WrapNotificationActionButton(ani_env* env,
     const std::shared_ptr<NotificationActionButton> &actionButton)
 {
-    if (actionButton == nullptr) {
-        ANS_LOGE("actionButton is null");
+    ANS_LOGD("WrapNotificationActionButton call");
+    if (env == nullptr || actionButton == nullptr) {
+        ANS_LOGE("WrapNotificationActionButton failed, has nullptr");
         return nullptr;
     }
     ani_object iconButtonObject = nullptr;
@@ -169,50 +231,56 @@ ani_object WrapNotificationActionButton(ani_env* env,
     ani_string stringValue = nullptr;
     if (!CreateClassObjByClassName(env,
         "Lnotification/notificationActionButton/NotificationActionButtonInner;", iconButtonCls, iconButtonObject)) {
+        ANS_LOGE("WrapNotificationActionButton : CreateClassObjByClassName failed");
         deletePointOfWrapNotificationActionButton(iconButtonObject, iconButtonCls, stringValue);
         return nullptr;
     }
     if (!SetNotificationActionButtonByRequiredParameter(env, iconButtonCls, iconButtonObject, actionButton)) {
+        ANS_LOGE("WrapNotificationActionButton : SetRequiredParameter failed");
         deletePointOfWrapNotificationActionButton(iconButtonObject, iconButtonCls, stringValue);
         return nullptr;
     }
     SetNotificationActionButtonByOptionalParameter(env, iconButtonCls, iconButtonObject, actionButton);
+    ANS_LOGE("WrapNotificationActionButton end");
     return iconButtonObject;
 }
 
 ani_status GetNotificationActionButtonArray(ani_env *env, ani_object param,
     const char *name, std::vector<std::shared_ptr<NotificationActionButton>> &res)
 {
+    ANS_LOGD("GetActionButtonArray call");
+    if (env == nullptr || param == nullptr || name == nullptr) {
+        ANS_LOGE("GetActionButtonArray failed, has nullptr");
+        return ANI_ERROR;
+    }
     ani_ref arrayObj = nullptr;
     ani_boolean isUndefined = true;
     ani_status status;
     ani_double length;
     StsActionButton actionButton;
     if ((status = GetPropertyRef(env, param, name, isUndefined, arrayObj)) != ANI_OK || isUndefined == ANI_TRUE) {
-        ANS_LOGI("status : %{public}d , %{public}s :  may be undefined", status, name);
+        ANS_LOGE("GetActionButtonArray: GetPropertyRef name = %{public}s, status = %{public}d", name, status);
         deletePoint(arrayObj);
         return ANI_INVALID_ARGS;
     }
-    status = GetPropertyDouble(env, static_cast<ani_object>(arrayObj), "length", isUndefined, length);
-    if (status != ANI_OK) {
-        ANS_LOGI("status : %{public}d", status);
+    if (ANI_OK!= (status = GetPropertyDouble(env, static_cast<ani_object>(arrayObj), "length", isUndefined, length))) {
+        ANS_LOGE("GetActionButtonArray: GetPropertyDouble name = %{public}s, status = %{public}d", name, status);
         deletePoint(arrayObj);
         return status;
     }
-
-    for (int i = 0; i < static_cast<int>(length); i++) {
+    for (int dex = 0; dex < static_cast<int>(length); dex++) {
         ani_ref buttonRef;
-        status = env->Object_CallMethodByName_Ref(static_cast<ani_object>(arrayObj),
-            "$_get", "I:Lstd/core/Object;", &buttonRef, (ani_int)i);
-        if (status != ANI_OK) {
+        if (ANI_OK != (status = env->Object_CallMethodByName_Ref(static_cast<ani_object>(arrayObj),
+            "$_get", "I:Lstd/core/Object;", &buttonRef, (ani_int)dex))) {
             deletePoint(arrayObj);
             deletePoint(buttonRef);
-            ANS_LOGI("status : %{public}d, index: %{public}d", status, i);
+            ANS_LOGE("GetActionButtonArray: get ref failed, status = %{public}d, index = %{public}d", status, dex);
             return status;
         }
-        status = UnwrapNotificationActionButton(env, static_cast<ani_object>(buttonRef), actionButton);
-        if (status != ANI_OK) {
-            ANS_LOGI("ActionButton failed, index: %{public}d", i);
+        if (ANI_OK
+            != (status = UnwrapNotificationActionButton(env, static_cast<ani_object>(buttonRef), actionButton))) {
+            ANS_LOGE("GetActionButtonArray: UnwrapActionButton failed, status = %{public}d, index = %{public}d",
+                status, dex);
             deletePoint(arrayObj);
             deletePoint(buttonRef);
             return status;
@@ -224,19 +292,21 @@ ani_status GetNotificationActionButtonArray(ani_env *env, ani_object param,
             actionButton.userInput, actionButton.isContextual);
         res.push_back(button);
     }
+    ANS_LOGD("GetActionButtonArray end");
     return status;
 }
 
 ani_object GetAniArrayNotificationActionButton(ani_env* env,
     const std::vector<std::shared_ptr<NotificationActionButton>> &actionButtons)
 {
-    if (actionButtons.empty()) {
-        ANS_LOGE("actionButtons is empty");
+    ANS_LOGD("GetAniArrayActionButton call");
+    if (env == nullptr || actionButtons.empty()) {
+        ANS_LOGE("GetAniArrayActionButton failed, has nullptr");
         return nullptr;
     }
     ani_object arrayObj = newArrayClass(env, actionButtons.size());
     if (arrayObj == nullptr) {
-        ANS_LOGE("arrayObj is empty");
+        ANS_LOGE("GetAniArrayActionButton: arrayObj is nullptr");
         deletePoint(arrayObj);
         return nullptr;
     }
@@ -244,17 +314,20 @@ ani_object GetAniArrayNotificationActionButton(ani_env* env,
     for (auto &button : actionButtons) {
         ani_object item = WrapNotificationActionButton(env, button);
         if (item == nullptr) {
+            ANS_LOGE("GetAniArrayActionButton: item is nullptr");
             deletePoint(arrayObj);
             deletePoint(item);
             return nullptr;
         }
         if (ANI_OK != env->Object_CallMethodByName_Void(arrayObj, "$_set", "ILstd/core/Object;:V", index, item)) {
+            ANS_LOGE("GetAniArrayActionButton: Object_CallMethodByName_Void failed");
             deletePoint(arrayObj);
             deletePoint(item);
             return nullptr;
         }   
         index ++;
     }
+    ANS_LOGD("GetAniArrayActionButton end");
     return arrayObj;
 }
 }
